+++ TP_1/src/calculos.h
@@ -8,18 +8,18 @@
 #ifndef CALCULOS_H_
 #define CALCULOS_H_
 
-int descuentoDebitoA(float precioA);
-int descuentoDebitoL(float precioL);
+float descuentoDebitoA(float precioA);
+float descuentoDebitoL(float precioL);
 
-int interesCreditoA(float precioA);
-int interesCreditoL(float precioL);
+float interesCreditoA(float precioA);
+float interesCreditoL(float precioL);
 
-int bitcoinA(float precioA);
-int bitcoinL(float precioL);
+float bitcoinA(float precioA);
+float bitcoinL(float precioL);
 
-int precioUnitarioA(float precioA, float km);
-int precioUnitarioL(float precioL, float km);
+float precioUnitarioA(float precioA, float km);
+float precioUnitarioL(float precioL, float km);
 
-int diferenciaAerolineas(float precioA, float precioL);
+float diferenciaAerolineas(float precioA, float precioL);
 
 #endif /* CALCULOS_H_ */
